Use buffer to handle template runtime errors properly
This commit utilizes a buffer to handle template runtime errors. We
execute the template into a buffer instead of the http.ResponseWriter.
If we get any errors we call our serverError helper and return. If we do
not get any errors then we utilizes the buffer's WriteTo method and
write to the http.ResponseWriter.

This way we properly handle errors and
return the correct http status code if there is such an issue.