Create forms package to abstract out form and form error handling
This commit creates the forms package to abstract out the form and form
error handling in the createSnippets and craeteSnippetsForm handlers. It
It does this by creating a Form struct and FormErrors struct with
corresponding methods on each struct to appropriately handle form
validation and form error handling which easily extends to templates and
new form validation as the app grows.