Add logic to UserModel Get method
This commit adds the logic for the UserModel's Get method. Specifically,
this commit queries the database users table for the user based on their
id and then inserts the user's active value into the initialized User
struct with the corresponding Active struct field. The method then
returns this initialized User struct.