Use request context to add authenticate middleware
This commit utilizes the request context to create an authenticate
middleware function. The middleware function first checks the session
data to see if the user is authenticated. If the user is not
authenticated then it calls the next handler in the middleware chain as
normal.

Next, the middleware function pulls the users data from the "users"
table and checks if they are still active or if they are in the
database at all. If neither condition is true then we remove
authenticatedUserID session value and call the next handler in the
chain.

Otherwise, we know the request is coming from an activated,
authenticated user and so we create a new request context with a true
boolean under the contextKeyIsAuthenticated typed key and call the next
handler in the chain.

We also change the isAuthenticated helper function to call the request
context's value to pull the contextKeyIsAuthenticated key's boolean
value. If no value is present we assume they are not authenticated and
return false.