Add parallel testing
This commit adds parallel testing to the existing tests. Also, we
rebind the test case from sub-test loop to avoid concurrency bug when
running sub-tests (if we don't do this, by the time the goroutines for
the tests launches the tt variable in the loop iterator is already
evaluated and each goroutine for the sub-test would be using the last
loop value, the CET test).

Note, this commit also adds t.Logf for the sub-tests to showcase the
concurrency bug introduced when running sub-tests in parallel via
table-driven testing.