Refactor test server returned by constructor
This commit customizes the test server returned by the test server
constructor. Specifically, we create a cookiejar for the test client
to automatically store any cookies sent in an HTTP response so that we
can include them (if required) in any subsequent requests back to the
test server.

Also, we disable redirect-following (i.e. prevent the client from
following redirects) so that we can test the response for that specific
request.