Refactor newTestApp
This commit refactors the testutils_test constructor newTestApp()
function to return a full mocked application struct. The loggers are
mocked in that they print to io.Discard. We also utilize a templateCache
and a session manager with the settings as production. Last, we utilize
the mocked versions of SnippetModel and UserModel.