Create sql test utilities
This commit creates setup and teardown SQL scripts to ensure we use a
new test DB for each integration test and ensure we tear it down after
each integration test. The scripts are utilized in the newTestDB
constructor which connects to the test database, runs the setup SQL
script, and then returns an anonymous teardown function to be run
defer'd when executed during testing.